movie_recommender/requirements.txt
pandas
scikit-learn
numpy
requests
pyyaml
pillow
ttkbootstrap